getyttranscript/requirements.txt
azure-functions
youtube-transcript-api
selenium
webdriver_manager